requirements.txt
Flask==2.0.1
tensorflow==1.15.0
numpy==1.19.5
pillow==8.0.1
